+++ src/components/CaouselMain/carousellData.js
@@ -1,0 +1,97 @@
+export const responsive = {
+    superLargeDesktop: {
+      breakpoint: { max: 4000, min: 1024 },
+      items: 4,
+      slidesToSlide: 4,
+    },
+    desktop: {
+      breakpoint: { max: 1024, min: 800 },
+      items: 3,
+      slidesToSlide: 3,
+    },
+    tablet: {
+      breakpoint: { max: 800, min: 464 },
+      items: 2,
+      slidesToSlide: 2,
+    },
+    mobile: {
+      breakpoint: { max: 464, min: 0 },
+      items: 1,
+      slidesToSlide: 1,
+    }
+  };
+
+export const cardsData = [
+    {
+      id: 1,
+      imageUrl: "/public/Assets/images/grecia.png",
+      name: "Primera imagen",
+      city: "Greece"
+    },
+    {
+      id: 2,
+      imageUrl: "/public/Assets/images/Republicadominicana.png",
+      name: "Primera imagen",
+      city: "Dominican Republic"
+    },
+    {
+      id: 3,
+      imageUrl: "/public/Assets/images/Mexico.jpg",
+      name: "Primera imagen",
+      city: "Mexico"
+    },
+    {
+      id: 4,
+      imageUrl: "/public/Assets/images/Brasil.jpg",
+      name: "Primera imagen",
+      city: "Brasil"
+    },
+    {
+      id: 5,
+      imageUrl: "/public/Assets/images/BuenosAires.jpg",
+      name: "Primera imagen",
+      city: "Buenos Aires"
+    },
+    {
+      id: 6,
+      imageUrl: "/public/Assets/images/Disney.jpg",
+      name: "Primera imagen",
+      city: "Disneyland"
+    },
+    {
+      id: 7,
+      imageUrl: "/public/Assets/images/NewYork.jpg",
+      name: "Primera imagen",
+      city: "New York"
+    },
+    {
+      id: 8,
+      imageUrl: "/public/Assets/images/Francia.jpg",
+      name: "Primera imagen",
+      city: "Francia"
+    },
+    {
+      id: 9,
+      imageUrl: "/public/Assets/images/Copenhagen-Dinamarca.jpg",
+      name: "Primera imagen",
+      city: "Copenhagen"
+    },
+    {
+      id: 10,
+      imageUrl: "/public/Assets/images/India-TajMahal.jpg",
+      name: "Primera imagen",
+      city: "India"
+    },
+    {
+      id: 11,
+      imageUrl: "/public/Assets/images/Sevilla-España.jpg",
+      name: "Primera imagen",
+      city: "España"
+    },
+    {
+      id: 12,
+      imageUrl: "/public/Assets/images/Turquia.jpg",
+      name: "Primera imagen",
+      city: "Turquia"
+    },
+  ]
