+++ src/components/Navbar/NavbarMain.jsx
@@ -5,9 +5,6 @@ import Form from 'react-bootstrap/Form';
 import Nav from 'react-bootstrap/Nav';
 import Navbar from 'react-bootstrap/Navbar';
 
-
-
-
 const NavbarMain = () => {
     return (
         <Navbar expand="lg" className="bg-body-tertiary found" fixed="top">
@@ -17,7 +14,7 @@ const NavbarMain = () => {
                 <Navbar.Collapse id="navbarScroll">
                     <Nav className="me-auto my-2 my-lg-0" style={{ maxHeight: '100px' }} navbarScroll>
                         <Nav.Link href="#action1">Home</Nav.Link>
-                        <Nav.Link href="#action2">Cyties</Nav.Link>
+                        <Nav.Link href="#action2">Cities</Nav.Link>
                     </Nav>
                     <Form className="d-flex">
                         <Button variant="outline-info">Login</Button>
@@ -25,7 +22,6 @@ const NavbarMain = () => {
                 </Navbar.Collapse>
             </Container>
         </Navbar>
-
     )
 }
